+++ referenceframe/testfiles/example_gantry.xml
