+++ referenceframe/testfiles/kinematicsloop.json
