--- referenceframe/urdf/geometry.go
@@ -1,71 +1,0 @@
-package urdf
-
-import (
-	"encoding/xml"
-	"fmt"
-
-	"github.com/golang/geo/r3"
-	"github.com/pkg/errors"
-
-	"go.viam.com/rdk/spatialmath"
-	"go.viam.com/rdk/utils"
-)
-
-var errGeometryTypeUnsupported = errors.New("unsupported Geometry type")
-
-// collision is a struct which details the XML used in a URDF collision geometry.
-type collision struct {
-	XMLName  xml.Name `xml:"collision"`
-	Origin   *pose    `xml:"origin"`
-	Geometry struct {
-		XMLName xml.Name `xml:"geometry"`
-		Box     *box     `xml:"box,omitempty"`
-		Sphere  *sphere  `xml:"sphere,omitempty"`
-	} `xml:"geometry"`
-}
-
-type box struct {
-	XMLName xml.Name `xml:"box"`
-	Size    string   `xml:"size,attr"` // "x y z" format, in meters
-}
-
-type sphere struct {
-	XMLName xml.Name `xml:"sphere"`
-	Radius  float64  `xml:"radius,attr"` // in meters
-}
-
-func newCollision(g spatialmath.Geometry) (*collision, error) {
-	cfg, err := spatialmath.NewGeometryConfig(g)
-	if err != nil {
-		return nil, err
-	}
-	urdf := &collision{
-		Origin: newPose(g.Pose()),
-	}
-	//nolint:exhaustive
-	switch cfg.Type {
-	case spatialmath.BoxType:
-		urdf.Geometry.Box = &box{Size: fmt.Sprintf("%f %f %f", utils.MMToMeters(cfg.X), utils.MMToMeters(cfg.Y), utils.MMToMeters(cfg.Z))}
-	case spatialmath.SphereType:
-		urdf.Geometry.Sphere = &sphere{Radius: utils.MMToMeters(cfg.R)}
-	default:
-		return nil, fmt.Errorf("%w %s", errGeometryTypeUnsupported, fmt.Sprintf("%T", cfg.Type))
-	}
-	return urdf, nil
-}
-
-func (c *collision) toGeometry() (spatialmath.Geometry, error) {
-	switch {
-	case c.Geometry.Box != nil:
-		dims := spaceDelimitedStringToFloatSlice(c.Geometry.Box.Size)
-		return spatialmath.NewBox(
-			c.Origin.Parse(),
-			r3.Vector{X: utils.MetersToMM(dims[0]), Y: utils.MetersToMM(dims[1]), Z: utils.MetersToMM(dims[2])},
-			"",
-		)
-	case c.Geometry.Sphere != nil:
-		return spatialmath.NewSphere(c.Origin.Parse(), utils.MetersToMM(c.Geometry.Sphere.Radius), "")
-	default:
-		return nil, errors.New("couldn't parse xml: no geometry defined")
-	}
-}
